[HUDI-6606] Use record level index with SQL equality queries
